Created JavaScript modules for every script features
