+++ index.css
@@ -15,7 +15,7 @@ nav {
   align-self: stretch;
   display: flex;
   justify-content: space-between;
-  padding: 12px 24px;
+  padding: 12px 1%;
   box-shadow: 0 0.5rem 1rem rgb(0 0 0 / 15%);
   align-items: center;
   position: fixed;
@@ -73,8 +73,6 @@ nav a {
 
 section {
   margin-top: 88px;
-  padding: 0 24px;
-  max-width: calc(100% - 48px);
 }
 
 .background {
@@ -134,7 +132,6 @@ span {
   text-align: center;
   display: flex;
   flex-direction: column;
-  margin: 20px;
 }
 
 #book-submit > * {
