+++ index.html
@@ -5,6 +5,7 @@
   <meta http-equiv="X-UA-Compatible" content="IE=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1.0">
 
+  <script type="module" src="./modules/book-functions.js"></script>
   <link rel="stylesheet" href="index.css">
 
   <title>Awesome Books</title>
@@ -62,6 +63,6 @@
     </div>
   </section>
 
-  <script src="./index.js"></script>
+  <script type="module" src="./index.js"></script>
 </body>
 </html>
